Correct timing note for stock TensorBoard
On re-timing, this takes only around 50 seconds, not 750, which is a big
difference. The original 750-second timing for stock TensorBoard was
inferred from this run of `tensorboard-memory-profiling`:
<https://gist.github.com/wchargin/f903494eb10a983aa7693f8fe0debc79>

My estimate was due to the log entry at 752.12 seconds, which was the
last time that `TensorEvent` values were allocated; `TensorEvent`s
should only be allocated while data is still loading. Furthermore, there
is no appreciable change in the allocation pattern of `TensorEvent`s up
until 752 seconds, even though we now expect that loading actually
finished around 50 seconds in.

The memory profiling overhead itself likely contributes somewhat to the
difference, but more importantly there was an active frontend with
auto-reload enabled (by design), causing lots of locking contention.
It’s still not clear to me why the difference was so large, but in any
case this was never a fair comparison, and the new results are clear.

For reference, the revised methodology is to run

    stdbuf -o0 -e0 tensorboard \
        --port 0 \
        --host localhost \
        --reload_task blocking \
        --reload_interval 0 \
        --logdir /path/to/logs \
        2>&1 |
    ts -s '[%.s]'

with no incoming connections to the server, and report the timestamp at
which the “TensorBoard [version] at [URL]” log line appears.